Refactor import mapping and file scanning logic
The change was aimed at improving the efficiency and readability of code related to import mapping and file scanning. The config class was simplified by removing 'compatibleExtensions' array and moving it into a constant as these values do not change. The logic to replace import paths was adjusted to also handle asset imports such as .png, .jpg, pdf files which were being ignored in the past. The scanner class was also heavily refactored to better handle directory parsing and more efficiently link imports to their corresponding code files. This was necessary to improve the code maintainability and performance of the application.